requirements.txt
google-generativeai
streamlit
ipykernel